Quiz about Python written in Python
This Program, written in Python for the Udacity Introduction to Programming Nano Degree, asks the user questions about the Python language. It has three levels easy, medium and hard. The user selects the difficulty level and the number of attempts and must fill in the gaps in a quiz question with the correct answers to complete the quiz.